refactor: ItemType 패키지 변경
